rename enhanced clinical card
