+++ src/packages/@ncigdc/modern_components/ClinicalAnalysis/ClinicalVariableCard/SharedVariableCard.js
